+++ NEWS.md
@@ -4,6 +4,7 @@
 
 ## New features / user-visible changes
 
+- New `add_github_action()` and `add_gitlab_ci()` helpers generate minimal deployment CI for fresh `{golem}` apps.
 - The `add_dockerfile_with_renv_*` function now generates a multi-stage Dockerfile by default (use `single_file = FALSE` to retain the previous behavior).
 - The `add_dockerfile_with_renv_*` function now creates a Dockerfile that sets `golem.app.prod = TRUE` by default (use `set_golem.app.prod = FALSE` to retain the previous behavior).
 - Print functions have be reworked standardized using the `{cli}` package (@ilyaZar, #89)